+++ apps/tree-shaking-tester/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @repo/tree-shaking-tester
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @giosg-design-system/icons@1.1.0
+
 ## 0.0.3
 
 ### Patch Changes